# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_NR3101_go-ecom-project_internal_dto.AddToCartRequest:
    properties:
      product_id:
        type: integer
      quantity:
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
  github_com_NR3101_go-ecom-project_internal_dto.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.UserResponse'
    type: object
  github_com_NR3101_go-ecom-project_internal_dto.CartItemResponse:
    properties:
      id:
        type: integer
      product:
        $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.ProductResponse'
      quantity:
        type: integer
      subtotal:
        type: number
    type: object
  github_com_NR3101_go-ecom-project_internal_dto.CartResponse:
    properties:
      cart_items:
        items:
          $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.CartItemResponse'
        type: array
      id:
        type: integer
      total:
        type: number
      user_id:
        type: integer
    type: object
  github_com_NR3101_go-ecom-project_internal_dto.CategoryResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
    type: object
...
  github_com_NR3101_go-ecom-project_internal_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  github_com_NR3101_go-ecom-project_internal_dto.ProductImageResponse:
    properties:
      alt_text:
        type: string
      id:
        type: integer
      is_primary:
        type: boolean
      url:
        type: string
    type: object
  github_com_NR3101_go-ecom-project_internal_dto.ProductResponse:
    properties:
      category:
        $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.CategoryResponse'
      category_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.ProductImageResponse'
        type: array
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock:
        type: integer
    type: object
  github_com_NR3101_go-ecom-project_internal_dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  github_com_NR3101_go-ecom-project_internal_dto.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  github_com_NR3101_go-ecom-project_internal_dto.UpdateCartItemRequest:
    properties:
      quantity:
        minimum: 1
        type: integer
    required:
    - quantity
    type: object
...
  github_com_NR3101_go-ecom-project_internal_dto.UserResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_name:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
  github_com_NR3101_go-ecom-project_internal_utils.PaginatedResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.PaginationMeta'
      success:
        type: boolean
    type: object
  github_com_NR3101_go-ecom-project_internal_utils.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_NR3101_go-ecom-project_internal_utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: no-email@enail.com
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for an e-commerce application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swaager.io/terms/
  title: E-commerce API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password
      parameters:
      - description: User sign-in data
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User logged in successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.AuthResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      summary: Login user
      tags:
      - Authentication
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user and invalidate refresh token
      parameters:
      - description: Refresh token to invalidate
        in: body
        name: logoutRequest
        required: true
        schema:
          $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User logged out successfully
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "500":
          description: Failed to logout user
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      summary: Logout user
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using refresh token
      parameters:
      - description: Refresh token data
        in: body
        name: refreshTokenRequest
        required: true
        schema:
          $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.AuthResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      summary: Refresh access token
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: User registration data
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.AuthResponse'
              type: object
        "400":
          description: Invalid request body or user already exists
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      summary: Register a new user
      tags:
      - Authentication
  /cart:
    get:
      consumes:
      - application/json
      description: Retrieve the shopping cart of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Cart retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.CartResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "404":
          description: Cart not found
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Get user's cart
      tags:
      - Cart
  /cart/items:
    post:
      consumes:
      - application/json
      description: Add a product to the authenticated user's cart
      parameters:
      - description: Add to cart data
        in: body
        name: addToCartRequest
        required: true
        schema:
          $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product added to cart successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.CartResponse'
              type: object
        "400":
          description: Invalid request body, product not found, or not enough stock
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - Cart
  /cart/items/{itemId}:
    delete:
      consumes:
      - application/json
      description: Remove a specific item from the cart
      parameters:
      - description: Cart Item ID
        in: path
        name: itemId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cart item removed successfully
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "400":
          description: Invalid cart item ID or cart item not found
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: Update the quantity of a specific item in the cart
      parameters:
      - description: Cart Item ID
        in: path
        name: itemId
        required: true
        type: integer
      - description: Update cart item data
        in: body
        name: updateCartItemRequest
        required: true
        schema:
          $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Cart item updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_dto.CartResponse'
              type: object
        "400":
          description: Invalid request body, cart item ID, or not enough stock
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Update cart item quantity
      tags:
      - Cart
...
  /orders:
    get:
      consumes:
      - application/json
      description: Retrieve all orders of the authenticated user with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Orders retrieved successfully
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Get all orders
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Create a new order from the user's cart
      produces:
      - application/json
      responses:
        "200":
          description: Order created successfully
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "400":
          description: Failed to create order (empty cart or insufficient stock)
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Create an order
      tags:
      - Orders
  /orders/{orderId}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific order by its ID
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order retrieved successfully
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "400":
          description: Invalid order ID
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/github_com_NR3101_go-ecom-project_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - Orders
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {token}" to authenticate requests
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"